2026-03-11 Fred Gleason <fredg@paravelsystems.com> 	* Removed Secret Rabbit Code dependency.

+++ link_common.sh
@@ -2,9 +2,9 @@
 
 # link_common.sh
 #
-#  Link common sources
+#  Link/Unlink common sources
 #
-#   (C) Copyright 2018 Fred Gleason <fredg@paravelsystems.com>
+#   (C) Copyright 2018-2026 Fred Gleason <fredg@paravelsystems.com>
 #
 #   This program is free software; you can redistribute it and/or modify
 #   it under the terms of the GNU General Public License version 2 as
@@ -20,24 +20,38 @@
 #   Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 #
 
+if test $# -ne 2 ; then
+    echo "link_common.sh <src-dir> link|unlink"
+    exit 1
+fi
+
 DESTDIR=$1
+OPERATION=$2
 
 rm -f src/$DESTDIR/alert.cpp
-ln -s ../../src/common/alert.cpp src/$DESTDIR/alert.cpp
 rm -f src/$DESTDIR/alert.h
-ln -s ../../src/common/alert.h src/$DESTDIR/alert.h
+if test $OPERATION = "link" ; then
+  ln -s ../../src/common/alert.cpp src/$DESTDIR/alert.cpp
+  ln -s ../../src/common/alert.h src/$DESTDIR/alert.h
+fi
 
 rm -f src/$DESTDIR/cmdswitch.cpp
-ln -s ../../src/common/cmdswitch.cpp src/$DESTDIR/cmdswitch.cpp
 rm -f src/$DESTDIR/cmdswitch.h
-ln -s ../../src/common/cmdswitch.h src/$DESTDIR/cmdswitch.h
+if test $OPERATION = "link" ; then
+  ln -s ../../src/common/cmdswitch.cpp src/$DESTDIR/cmdswitch.cpp
+  ln -s ../../src/common/cmdswitch.h src/$DESTDIR/cmdswitch.h
+fi
 
 rm -f src/$DESTDIR/config.cpp
-ln -s ../../src/common/config.cpp src/$DESTDIR/config.cpp
 rm -f src/$DESTDIR/config.h
-ln -s ../../src/common/config.h src/$DESTDIR/config.h
+if test $OPERATION = "link" ; then
+  ln -s ../../src/common/config.cpp src/$DESTDIR/config.cpp
+  ln -s ../../src/common/config.h src/$DESTDIR/config.h
+fi
 
 rm -f src/$DESTDIR/profile.cpp
-ln -s ../../src/common/profile.cpp src/$DESTDIR/profile.cpp
 rm -f src/$DESTDIR/profile.h
-ln -s ../../src/common/profile.h src/$DESTDIR/profile.h
+if test $OPERATION = "link" ; then
+  ln -s ../../src/common/profile.cpp src/$DESTDIR/profile.cpp
+  ln -s ../../src/common/profile.h src/$DESTDIR/profile.h
+fi
